[refactor] todoUpdate 기능 분리
