+++ 01.javascript/todoapp/assets/js/pages/delete/TodoDelete.js
@@ -1,5 +1,3 @@
-// /todolist/{_id}
-
 const TodoDelete = async (_id) => {
   console.log(_id);
   try {
